Rename lru.py to cache_ring.py and likewise with ring.c/h
