+++ relstorage/cache/cache_ring.h
